backend/analytics-ms/requirements.txt
fastapi
uvicorn[standard]
pymongo
dnspython
kafka-python
